+++ packages/com.yohawing.mmd-loader/Tests/EditMode/MmdAssetImporterTests.DragAndDrop.cs
@@ -23,43 +23,25 @@ namespace Mmd.Tests
     public sealed partial class MmdAssetImporterTests
     {
         [Test]
-        public void ImportedMmdAssetsUseLightweightCustomInspectors()
+        public void ImportedMmdAssetsUseScriptedImporterEditors()
         {
             CopyFixtureToAssetDatabase("test_1bone_cube.pmx", TempPmxPath);
             CopyFixtureToAssetDatabase("test_1bone_cube_motion.vmd", TempVmdPath);
-            MmdPmxAsset pmxAsset = AssetDatabase.LoadAssetAtPath<MmdPmxAsset>(TempPmxPath);
-            MmdVmdAsset vmdAsset = AssetDatabase.LoadAssetAtPath<MmdVmdAsset>(TempVmdPath);
             AssetImporter pmxImporter = AssetImporter.GetAtPath(TempPmxPath);
             AssetImporter vmdImporter = AssetImporter.GetAtPath(TempVmdPath);
-            UnityEditor.Editor? pmxEditor = null;
-            UnityEditor.Editor? vmdEditor = null;
             UnityEditor.Editor? pmxImporterEditor = null;
             UnityEditor.Editor? vmdImporterEditor = null;
 
             try
             {
-                pmxEditor = UnityEditor.Editor.CreateEditor(pmxAsset);
-                vmdEditor = UnityEditor.Editor.CreateEditor(vmdAsset);
                 pmxImporterEditor = UnityEditor.Editor.CreateEditor(pmxImporter);
                 vmdImporterEditor = UnityEditor.Editor.CreateEditor(vmdImporter);
 
-                Assert.That(pmxEditor, Is.TypeOf<MmdPmxAssetEditor>());
-                Assert.That(vmdEditor, Is.TypeOf<MmdVmdAssetEditor>());
                 Assert.That(pmxImporterEditor, Is.TypeOf<MmdPmxScriptedImporterEditor>());
                 Assert.That(vmdImporterEditor, Is.TypeOf<MmdVmdScriptedImporterEditor>());
             }
             finally
             {
-                if (pmxEditor != null)
-                {
-                    Object.DestroyImmediate(pmxEditor);
-                }
-
-                if (vmdEditor != null)
-                {
-                    Object.DestroyImmediate(vmdEditor);
-                }
-
                 if (pmxImporterEditor != null)
                 {
                     Object.DestroyImmediate(pmxImporterEditor);